Added logic to create dependencies #3982
While calling the jenkins plugin (Bayesian Analysis), 2 new mvn
commands were added for the direct & transitive dependency files.

The jenkins API will be calling the Analytics Server API. This
changes will make sure that when the call lands on to the scan API,
the API doesnt need to clone and redo all the logic to get the
dependencies.

Resolves https://github.com/openshiftio/openshift.io/issues/3982