Changed the command to generate dependencies #4361
The previous command 'resolve' was not working as expected in cases which
had multi maven modules.Hence had to be changed to 'collect'.

The 'resolve' command was failing in some scenarios where the artifact jars
were hosted on a private server and aren't downloadable without credentials.

The command 'collect' relies only on pom files to generate the dependency tree,
which allows the processing without the need to download the jars.
More Details: https://maven.apache.org/plugins/maven-dependency-plugin/collect-mojo.html.

Resolves: openshiftio/openshift.io#4361